% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACCR.R
\name{ACCR}
\alias{ACCR}
\title{ACCR function}
\usage{
ACCR(alpha_true, alpha_estimated)
}
\arguments{
\item{alpha_true}{The true attribute pattern of the examinee.}

\item{alpha_estimated}{The estimated attribute pattern of the examinee.}
}
\description{
The attribute correct classification rate measured the individual attribute recovery.
}
\details{
K is the number of attributes.
}
