% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EAP_estimation.R
\name{EAP_estimation}
\alias{EAP_estimation}
\title{EAP_estimation function}
\usage{
EAP_estimation(itembank, response, K)
}
\arguments{
\item{itembank}{the itembank including item_id, slip, guess, 4 attributes, content area.}

\item{response}{the responses of one examinee.}

\item{K}{the number of attributes.}
}
\description{
This function allows you to get the Expectation A Posterior Estimation.
}
\details{
prior is the uniform distribution.
}
