% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETA.R
\name{ETA}
\alias{ETA}
\title{ETA function}
\usage{
ETA(N, alpha, Q)
}
\arguments{
\item{N}{the examinee sample size.}

\item{alpha}{attribute pattern for one examinee or patterns for more than one examinee.}

\item{Q}{the Q matrix for all answered items.}
}
\value{
eta 1 indicates the required attributes are mastered, 0 otherwise.
}
\description{
This function allows you to identify whether the i_th examinee possesses all the attributes of item j.
}
\details{
for one person, alpha & eta are vectors; for more than one person, alpha & eta are matrices.
}
