% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSDest.R
\name{JSDest}
\alias{JSDest}
\title{JSDest function}
\usage{
JSDest(K, unused_items)
}
\arguments{
\item{K}{the number of attributes.}

\item{unused_items}{the unused items from the whole itembank including item_id, slip, guess, 4 attributes, content area.}
}
\description{
This function allows you to calculate the Jensen-Shannon Divergence Index.
}
\details{
prior is the uniform distribution.
}
