% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCprob.R
\name{LCprob}
\alias{LCprob}
\title{LCprob function}
\usage{
LCprob(K, unused_items)
}
\arguments{
\item{K}{the number of attributes.}

\item{unused_items}{the unused items from the whole itembank including item_id, slip, guess, 4 attributes, content area.}
}
\description{
This function allows you to calculate the latent class probability.
}
\details{
success probabilities for each category for all latent classes.
}
