% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P_DIAN.R
\name{P_DINA}
\alias{P_DINA}
\title{P_DINA function}
\usage{
P_DINA(N, slip, guess, eta)
}
\arguments{
\item{N}{the examinee sample size.}

\item{slip}{slipping parameters for all answered items.}

\item{guess}{slipping parameters for all answered items.}

\item{eta}{a value of 1 if examinee i possesses all the skills required for item j, otherwise 0.}
}
\description{
This function allows you to calculate the probability of the correct answer on item j for examinee i.
}
\details{
for one person, eta & p are vectors; for more than one person, eta & p are matrices.
}
