% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCCR.R
\name{PCCR}
\alias{PCCR}
\title{PCCR function}
\usage{
PCCR(alpha_true, alpha_estimated)
}
\arguments{
\item{alpha_true}{The true attribute pattern of the examinee.}

\item{alpha_estimated}{The estimated attribute pattern of the examinee.}
}
\description{
The pattern correct classification rate measured the entire attribute profile recovery.
}
\details{
I is the indicator function.
}
