% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWKLest.R
\name{PWKLest}
\alias{PWKLest}
\title{PWKLest function}
\usage{
PWKLest(K, unused_items, KS_estimate)
}
\arguments{
\item{K}{the number of attributes.}

\item{unused_items}{the unused items from the whole itembank including item_id, slip, guess, 4 attributes, content area.}

\item{KS_estimation}{the current KS estimation results.}
}
\description{
This function allows you to calculate the Posterior-weighted Kullback-Leibler Information.
}
\details{
prior is the uniform distribution.
}
