% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_gen.R
\name{response_gen}
\alias{response_gen}
\title{response_gen function}
\usage{
response_gen(N, slip, guess, eta)
}
\arguments{
\item{N}{the examinee sample size.}

\item{slip}{slipping parameters for all answered items.}

\item{guess}{slipping parameters for all answered items.}

\item{eta}{a value of 1 if examinee i possesses all the skills required for item j, otherwise 0.}
}
\description{
This function allows you to generate response matrix based on the DINA model.
}
\details{
for one person, resp are vectors; for more than one person, resp is a matrix.
}
