% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xi_likelihood.R
\name{Xi_likelihood}
\alias{Xi_likelihood}
\title{Xi_likelihood function}
\usage{
Xi_likelihood(N, J, p, X)
}
\arguments{
\item{N}{the examinee sample size.}

\item{J}{the number of items.}

\item{p}{the probability of the correct answer on item j for examinee i.}

\item{X}{the responses of one examinee or examinees.}
}
\description{
This function allows you to calculate the conditional joint maximum likelihood estimation.
}
\details{
for one person, p & X are vectors; for more than one person, p & X are matrices.
}
